requirements.txt
tensorflow
pillow
numpy
tk
streamlit 